+++ libs/link/component/src/lib/link-component.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { LinkProps } from '@digital-www-pwa/link-types';
-import { useTopNavContext } from '@digital-www-pwa/top-nav-context';
+import { useTopNavContext } from '@digital-www-pwa/providers';
 import { useRouter } from 'next/navigation';
 
 export function Link({ to, className, children }: LinkProps) {
