+++ libs/providers/src/index.ts
@@ -1,3 +1,4 @@
 // Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
 
 export * from './lib/FeedProvider';
+export * from './lib/TopNavProvider';
