--- libs/top-nav/context/README.md
@@ -1,7 +1,0 @@
-# top-nav-context
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test top-nav-context` to execute the unit tests via [Jest](https://jestjs.io).
