--- libs/top-nav/context/jest.config.ts
@@ -1,10 +1,0 @@
-export default {
-  displayName: 'top-nav-context',
-  preset: '../../../jest.preset.js',
-  transform: {
-    '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nx/react/plugins/jest',
-    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nx/react/babel'] }],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/top-nav/context',
-};
