--- libs/top-nav/context/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "top-nav-context",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/top-nav/context/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project top-nav-context --web",
-  "targets": {}
-}
