--- libs/top-nav/context/src/index.ts
@@ -1,3 +1,0 @@
-// Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
-
-export * from './lib/top-nav-context';
