--- libs/top-nav/context/src/lib/top-nav-context.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import TopNavContext from './top-nav-context';
-
-describe('TopNavContext', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<TopNavContext />);
-    expect(baseElement).toBeTruthy();
-  });
-});
