--- libs/top-nav/provider/README.md
@@ -1,7 +1,0 @@
-# top-nav-provider
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test top-nav-provider` to execute the unit tests via [Jest](https://jestjs.io).
