--- libs/top-nav/provider/jest.config.ts
@@ -1,10 +1,0 @@
-export default {
-  displayName: 'top-nav-provider',
-  preset: '../../../jest.preset.js',
-  transform: {
-    '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nx/react/plugins/jest',
-    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nx/react/babel'] }],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/top-nav/provider',
-};
