--- libs/top-nav/provider/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "top-nav-provider",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/top-nav/provider/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project top-nav-provider --web",
-  "targets": {}
-}
