--- libs/top-nav/provider/src/lib/top-nav-provider.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import TopNavProvider from './top-nav-provider';
-
-describe('TopNavProvider', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<TopNavProvider />);
-    expect(baseElement).toBeTruthy();
-  });
-});
