--- libs/top-nav/provider/src/server.ts
@@ -1,2 +1,0 @@
-// Use this file to export React server components
-export * from './lib/hello-server';
