--- libs/top-nav/types/README.md
@@ -1,3 +1,0 @@
-# top-nav-types
-
-This library was generated with [Nx](https://nx.dev).
