--- libs/top-nav/types/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@digital-www-pwa/top-nav-types",
-  "version": "0.0.1",
-  "dependencies": {},
-  "private": true
-}
