--- libs/top-nav/types/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "top-nav-types",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/top-nav/types/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project event-types --web",
-  "targets": {}
-}
