+++ libs/types/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/feed';
 export * from './lib/event';
+export * from './lib/top-nav';
