+++ tsconfig.base.json
@@ -29,11 +29,6 @@
       "@digital-www-pwa/top-nav-component": [
         "libs/top-nav/component/src/index.ts"
       ],
-      "@digital-www-pwa/top-nav-context": ["libs/top-nav/context/src/index.ts"],
-      "@digital-www-pwa/top-nav-provider": [
-        "libs/top-nav/provider/src/index.ts"
-      ],
-      "@digital-www-pwa/top-nav-types": ["libs/top-nav/types/src/index.ts"],
       "@digital-www-pwa/types": ["libs/types/src/index.ts"],
       "@digital-www-pwa/utils": ["libs/utils/src/index.ts"]
     }
